Extract `module` script plugin
